requirements.txt
# Postopus Production Deployment Requirements
# For Render.com deployment

# Core FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database Dependencies
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0

# HTTP Client for VK API
httpx==0.25.2
aiohttp==3.9.1

# Data Processing & Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Utility Libraries
python-dotenv==1.0.0
click==8.1.7
pytz==2023.3

# Development Tools (optional, for debugging)
pytest==7.4.3
pytest-asyncio==0.21.1